requirements.txt
PlexAPI
requests
python-dotenv
